[BD-222] Adds functionality to highlight the selected date range.
Disallow the selection of a start date on day that has N/A.
Disallow the selection of a stop date on a day less than the current date.
When the user is selecting a stop date, if the selected date is the same
as the start date then reset the date selection range.